+++ components/sidebar-nav.tsx
@@ -2,7 +2,6 @@
 
 import React from 'react'
 import Image from 'next/image'
-import Link from 'next/link'
 import { useRouter } from 'next/navigation'
 import { UserMenu } from 'components/user-menu'
 import { motion } from 'framer-motion'
@@ -42,42 +41,10 @@ const SidebarNav = () => {
         <UserMenu user={user} />
       </div>
       <div className="flex justify-center px-2 py-1 text-base font-medium rounded-lg">
-        <p className="text-center">Active Billing Organization:</p>
-      </div>
-      <div className="flex justify-center px-2 py-1 text-base font-medium rounded-lg">
-        <p>{activeBillingOrg ? activeBillingOrg.name : 'None'}</p>
+        <p>{activeBillingOrg ? activeBillingOrg.name : 'No org selected'}</p>
       </div>
       {/* Buttons section */}
       <nav className="flex flex-col space-y-1">
-        {/* Organizations Button */}
-        <Link
-          href="/orgs"
-          className="flex items-center px-2 py-1 text-base font-medium rounded-lg hover:bg-secondary"
-          style={{
-            color: 'var(--secondary-foreground)',
-            backgroundColor: 'var(--secondary)',
-            borderWidth: '1px',
-            borderColor: 'var(--border)',
-            borderRadius: 'var(--radius)',
-          }}
-        >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            fill="none"
-            viewBox="0 0 24 24"
-            strokeWidth="1.5"
-            stroke="currentColor"
-            className="w-6 h-6"
-          >
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              d="M2.25 21h19.5m-18-18v18m10.5-18v18m6-13.5V21M6.75 6.75h.75m-.75 3h.75m-.75 3h.75m3-6h.75m-.75 3h.75m-.75 3h.75M6.75 21v-3.375c0-.621.504-1.125 1.125-1.125h2.25c.621 0 1.125.504 1.125 1.125V21M3 3h12m-.75 4.5H21m-3.75 3.75h.008v.008h-.008v-.008Zm0 3h.008v.008h-.008v-.008Zm0 3h.008v.008h-.008v-.008Z"
-            />
-          </svg>
-          <span className="ml-3">Organizations</span>
-        </Link>
-
         {/* Projects Button */}
         <button
           className="flex items-center px-2 py-1 text-base font-medium rounded-lg hover:bg-secondary"